Do not defunct connection on TooLongFrameException
The previous implementation would defunct the connection and retry on
next host, which can be a problem since the query would ultimately be
tried on each host, raising this exception and closing the connection
each time.  Instead, keep the connection active, don't retry and
raise TooLongFrameException to the user.

Also replace AtomicReference implementation since we are assured that
decode will only be called one at a time on the same thread.

Change Decoder to only check version on first frame.

Change DecoderForStreamIdSize to not validate opcode until after the
full frame has been received.

Both these changes allow the parent LengthFieldBasedNameDecoder to skip
bytes when MAX_FRAME_LIMIT is exceeded properly and allow the Decoder to
properly proceed when skipping the rest of the data as part of consuming
the frame.